pylintrc: re-enable too-many-lines, final patch
To get under the 1000 line limit, move ycheck unit
tests into their own directory, and split-up into
sub-files as necessary.

Finally, enable the too-many-lines check in pylintrc.
