+++ pylintrc
@@ -30,5 +30,4 @@ disable=
  too-many-arguments,
  too-many-branches,
  too-many-instance-attributes,
- too-many-lines,
  too-many-locals,
