+++ tests/unit/ycheck/test_scenarios.py
@@ -13,7 +13,8 @@ from hotsos.core.ycheck.engine.properties.conclusions import (
     YPropertyConclusion,
 )
 
-from . import utils
+from .. import utils
+from . import test_scenarios_data as test_data
 
 
 class TestProperty():
@@ -39,555 +40,11 @@ class TestConfig(IniConfigBase):
     """ Test config """
 
 
-YDEF_NESTED_LOGIC = """
-checks:
-  isTrue:
-    requires:
-      and:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-        - property:
-            path: tests.unit.test_ycheck_scenarios.TestProperty.always_false
-            ops: [[not_]]
-  isalsoTrue:
-    requires:
-      or:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_false
-  isstillTrue:
-    requires:
-      and:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-      or:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_false
-  isnotTrue:
-    requires:
-      and:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-      or:
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_false
-        - property: tests.unit.test_ycheck_scenarios.TestProperty.always_false
-conclusions:
-  conc1:
-    decision:
-      and:
-        - isTrue
-        - isalsoTrue
-      or:
-        - isstillTrue
-    raises:
-      type: IssueTypeBase
-      message: conc1
-  conc2:
-    decision:
-      and:
-        - isTrue
-        - isnotTrue
-      or:
-        - isalsoTrue
-    raises:
-      type: IssueTypeBase
-      message: conc2
-  conc3:
-    decision:
-      not:
-        - isnotTrue
-      or:
-        - isalsoTrue
-    raises:
-      type: IssueTypeBase
-      message: conc3
-"""
-
-
-CONCLUSION_PRIORITY_1 = """
-checks:
-  testcheck:
-    property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-conclusions:
-  conc1:
-    priority: 1
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc1
-  conc2:
-    priority: 2
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc2
-  conc3:
-    priority: 3
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc3
-"""
-
-CONCLUSION_PRIORITY_2 = """
-checks:
-  testcheck:
-    property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-conclusions:
-  conc1:
-    priority: 1
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc1
-  conc2:
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc2
-  conc3:
-    decision:
-      - testcheck
-    raises:
-      type: IssueTypeBase
-      message: conc3
-"""
-
-
-YAML_DEF_REQUIRES_MAPPED = """
-checks:
-  is_exists_mapped:
-    systemd: nova-compute
-  is_exists_unmapped:
-    requires:
-      systemd: nova-compute
-conclusions:
-"""
-
-SCENARIO_W_EXPR_LIST = r"""
-input:
-  path: {path}
-checks:
-  listsearch1:
-    expr: ['hello y', 'hello x']
-  listsearch2:
-    search: ['hello y', 'hello x']
-  listsearch3:
-    search:
-       expr: ['hello y', 'hello x']
-conclusions:
-  listsearch1worked:
-    decision: listsearch1
-    raises:
-      type: SystemWarning
-      message: yay list search
-  listsearch2worked:
-    decision: listsearch2
-    raises:
-      type: SystemWarning
-      message: yay list search
-  listsearch3worked:
-    decision: listsearch3
-    raises:
-      type: SystemWarning
-      message: yay list search
-"""  # noqa
-
-
-SCENARIO_W_SEQ_SEARCH = r"""
-input:
-  path: {path}
-checks:
-  seqsearch1:
-    start: "it's the start"
-    body: ".+"
-    end: "it's the end"
-  seqsearch2:
-    start: "it's the start"
-    end: "it's the end"
-conclusions:
-  seqsearchworked:
-    decision:
-      - seqsearch1
-      - seqsearch2
-    raises:
-      type: SystemWarning
-      message: yay seq searches worked!
-"""  # noqa
-
-
-SCENARIO_W_ERROR = r"""
-scenarioA:
-  checks:
-    property_no_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-  conclusions:
-    c1:
-      decision: property_no_error
-      raises:
-        type: SystemWarning
-        message: foo
-scenarioB:
-  checks:
-    property_w_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.i_dont_exist
-  conclusions:
-    c1:
-      decision: property_w_error
-      raises:
-        type: SystemWarning
-        message: foo
-"""  # noqa
-
-
-CONCLUSION_W_INVALID_BUG_RAISES = r"""
-scenarioA:
-  checks:
-    property_no_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-  conclusions:
-    c1:
-      decision: property_no_error
-      raises:
-        type: SystemWarning
-        bug-id: 1234
-        message: foo
-scenarioB:
-  checks:
-    property_w_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-  conclusions:
-    c1:
-      decision: property_no_error
-      raises:
-        type: LaunchpadBug
-        message: foo
-scenarioC:
-  checks:
-    property_w_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-  conclusions:
-    c1:
-      decision: property_no_error
-      raises:
-        type: UbuntuCVE
-        message: foo
-scenarioD:
-  checks:
-    property_w_error:
-      property: tests.unit.test_ycheck_scenarios.TestProperty.always_true
-  conclusions:
-    c1:
-      decision: property_no_error
-      raises:
-        cve-id: 123
-"""  # noqa
-
-
-SCENARIO_CHECKS = r"""
-checks:
-  logmatch:
-    input:
-      path: foo.log
-    expr: '(\d{4}-\d{2}-\d{2}) (\d{2}:\d{2}:\d{2}\.\d{3}) (\S+) \S+'
-    constraints:
-      min-results: 3
-      search-period-hours: 24
-      search-result-age-hours: 48
-  property_true_shortform:
-    requires:
-      property:
-        path: hotsos.core.plugins.system.system.SystemBase.virtualisation_type
-  property_has_value_longform:
-    requires:
-      property:
-        path: hotsos.core.plugins.system.system.SystemBase.virtualisation_type
-        ops: [[eq, kvm], [truth], [not_], [not_]]
-  apt_pkg_exists:
-    requires:
-      apt: nova-compute
-  snap_pkg_exists:
-    requires:
-      snap: core20
-  service_exists_short:
-    requires:
-      systemd: nova-compute
-  service_exists_and_enabled:
-    requires:
-      systemd:
-        nova-compute: enabled
-  service_exists_not_enabled:
-    requires:
-      systemd:
-        nova-compute:
-          state: enabled
-          op: ne
-conclusions:
-  justlog:
-    priority: 1
-    decision: logmatch
-    raises:
-      type: SystemWarning
-      message: log matched {num} times ({group})
-      format-dict:
-        num: '@checks.logmatch.search.num_results'
-        group: '@checks.logmatch.search.results_group_2:comma_join'
-  logandsnap:
-    priority: 2
-    decision:
-      and:
-        - logmatch
-        - snap_pkg_exists
-    raises:
-      type: SystemWarning
-      message: log matched {num} times and snap exists
-      format-dict:
-        num: '@checks.logmatch.search.num_results'
-  logandsnapandservice:
-    priority: 3
-    decision:
-      and:
-        - logmatch
-        - snap_pkg_exists
-        - service_exists_short
-        - service_exists_and_enabled
-        - property_true_shortform
-        - property_has_value_longform
-    raises:
-      type: SystemWarning
-      message: log matched {num} times, snap and service exists
-      format-dict:
-        num: '@checks.logmatch.search.num_results'
-"""  # noqa
-
-
-CONFIG_SCENARIO = """
-checks:
-  cfg_is_bad:
-    config:
-      handler: tests.unit.test_ycheck_scenarios.TestConfig
-      path: test.conf
-      assertions:
-        - key: key1
-          section: DEFAULT
-          ops: [[lt, 102]]
-        - key: key1
-          section: DEFAULT
-          ops: [[gt, 100]]
-  cfg_is_bad2:
-    config:
-      handler: tests.unit.test_ycheck_scenarios.TestConfig
-      path: test.conf
-      assertions:
-        not:
-          - key: key1
-            section: DEFAULT
-            ops: [[lt, 103]]
-          - key: key1
-            section: DEFAULT
-            ops: [[gt, 101]]
-conclusions:
-  cfg_is_bad:
-    decision: cfg_is_bad
-    raises:
-      type: SystemWarning
-      message: cfg is bad
-  cfg_is_bad2:
-    decision: cfg_is_bad2
-    raises:
-      type: SystemWarning
-      message: cfg is bad2
-"""
-
-
-VARS = """
-vars:
-  foo: 1000
-  limit: 10
-  bar: "two"
-  frombadprop: '@tests.unit.idontexist'  # add to ensure lazy-loaded
-  fromprop: '@tests.unit.test_ycheck_scenarios.TestProperty.myattr'
-  fromfact: '@hotsos.core.host_helpers.systemd.ServiceFactory.start_time_secs:snapd'
-  fromfact2: '@hotsos.core.host_helpers.filestat.FileFactory.mtime:myfile.txt'
-  fromsysctl: '@hotsos.core.host_helpers.sysctl.SYSCtlFactory:net.core.somaxconn'
-  boolvar: false
-checks:
-  aptcheck:
-    apt: nova-compute
-  is_foo_lt:
-    varops: [[$foo], [lt, $limit]]
-  is_foo_gt:
-    varops: [[$foo], [gt, $limit]]
-  isbar:
-    varops: [[$bar], [ne, ""]]
-  fromprop:
-    varops: [[$fromprop], [eq, "123"]]
-  fromfact:
-    varops: [[$fromfact], [gt, 1644446300]]
-  fromfact2:
-    varops: [[$fromfact2], [eq, 0]]
-  fromsysctl:
-    varops: [[$fromsysctl], [eq, '4096']]
-  boolvar:
-    varops: [[$boolvar], [truth], [not_]]
-conclusions:
-  aptcheck:
-    decision: aptcheck
-    raises:
-      type: SystemWarning
-      message: "{name}={version}"
-      format-dict:
-        name: '@checks.aptcheck.requires.package'
-        version: '@checks.aptcheck.requires.version'
-  is_foo_gt:
-    decision: is_foo_gt
-    raises:
-      type: SystemWarning
-      message: it's foo gt! ({varname}={varval})
-      format-dict:
-        varname: '@checks.is_foo_gt.requires.input_ref'
-        varval: '@checks.is_foo_gt.requires.input_value'
-  is_foo_lt:
-    decision: is_foo_lt
-    raises:
-      type: SystemWarning
-      message: it's foo lt! ({varname}={varval})
-      format-dict:
-        varname: '@checks.is_foo_lt.requires.input_ref'
-        varval: '@checks.is_foo_lt.requires.input_value'
-  isbar:
-    decision: isbar
-    raises:
-      type: SystemWarning
-      message: it's bar! ({varname}={varval})
-      format-dict:
-        varname: '@checks.isbar.requires.input_ref'
-        varval: '@checks.isbar.requires.input_value'
-  fromprop:
-    decision:
-      and: [fromprop, boolvar, fromfact, fromfact2, fromsysctl]
-    raises:
-      type: SystemWarning
-      message: fromprop! ({varname}={varval})
-      format-dict:
-        varname: '@checks.fromprop.requires.input_ref'
-        varval: '@checks.fromprop.requires.input_value'
-"""  # noqa
-
-
-# this set of checks should cover the variations of logical op groupings that
-# will not process items beyond the first that returns False which is the
-# default for any AND operation since a single False makes the group result the
-# same.
-LOGIC_TEST = """
-vars:
-  # this one will resolve as False
-  v1: '@tests.unit.test_ycheck_scenarios.TestProperty.always_false'
-  # this one will raise an ImportError
-  v2: '@tests.unit.test_ycheck_scenarios.TestProperty.doesntexist'
-checks:
-  # the second item in each group must be one that if evaluated will raise an
-  # error.
-  chk_and:
-    and:
-      - varops: [[$v1], [truth]]
-      - varops: [[$v2], [not_]]
-  chk_nand:
-    nand:
-      - varops: [[$v1], [not_]]
-      - varops: [[$v2], [not_]]
-  chk_not:
-    not:
-      - varops: [[$v1], [not_]]
-      - varops: [[$v2], [not_]]
-  chk_default_and:
-    - varops: [[$v2], [not_]]
-    - varops: [[$v1], [truth]]
-conclusions:
-  conc1:
-    decision:
-       or:
-         - chk_and
-         - chk_nand
-         - chk_not
-         - chk_default_and
-    raises:
-      type: SystemWarning
-      message: >-
-        This should never get raised since all checks should be
-        returning False and the first false result in each check's logical
-        group should result in further items not being executed.
-"""
-
-NESTED_LOGIC_TEST_W_ISSUE = """
-vars:
-  bool_true: true
-checks:
-  chk_pass1:
-    and:
-      - varops: [[$bool_true], [truth]]
-      - not:
-          varops: [[$bool_true], [not_]]
-  chk_pass2:
-    or:
-      - and:
-          - varops: [[$bool_true], [truth]]
-          - varops: [[$bool_true], [not_]]
-      - and:
-          - varops: [[$bool_true], [truth]]
-          - varops: [[$bool_true], [truth]]
-      - varops: [[$bool_true], [truth]]
-  chk_pass3:
-    or:
-      - varops: [[$bool_true], [truth]]
-      - varops: [[$bool_true], [truth]]
-conclusions:
-  conc1:
-    decision: [chk_pass1, chk_pass2, chk_pass3]
-    raises:
-      type: SystemWarning
-      message:
-"""
-
-
-NESTED_LOGIC_TEST_NO_ISSUE = """
-vars:
-  bool_true: true
-checks:
-  chk_fail1:
-    and:
-      - varops: [[$bool_true], [truth]]
-      - not:
-          varops: [[$bool_true], [truth]]
-  chk_fail2:
-    or:
-      - and:
-          - varops: [[$bool_true], [not_]]
-          - varops: [[$bool_true], [not_]]
-      - and:
-          - varops: [[$bool_true], [not_]]
-          - varops: [[$bool_true], [not_]]
-      - varops: [[$bool_true], [not_]]
-conclusions:
-  conc1:
-    decision:
-      or: [chk_fail1, chk_fail2]
-    raises:
-      type: SystemWarning
-      message:
-"""
-
-DPKG_L = """
-ii  openssh-server                       1:8.2p1-4ubuntu0.4                                   amd64        secure shell (SSH) server, for secure access from remote machines
-"""  # noqa
-
-
 class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-public-methods
     """
     Tests scenarios functionality.
     """
-    @utils.init_test_scenario(SCENARIO_W_EXPR_LIST.
+    @utils.init_test_scenario(test_data.SCENARIO_W_EXPR_LIST.
                               format(path=os.path.basename('data.txt')))
     @utils.create_data_root({'data.txt': 'hello x\n'})
     @utils.global_search_context
@@ -603,7 +60,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
             msg = "yay list search"
             self.assertEqual(issue['message'], msg)
 
-    @utils.init_test_scenario(SCENARIO_W_SEQ_SEARCH.
+    @utils.init_test_scenario(test_data.SCENARIO_W_SEQ_SEARCH.
                               format(path=os.path.basename('data.txt')))
     @utils.create_data_root({'data.txt': ("blah blah\nit's the start\nblah "
                                           "blah\nit's the end")})
@@ -618,7 +75,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
             msg = "yay seq searches worked!"
             self.assertEqual(issue['message'], msg)
 
-    @utils.init_test_scenario(SCENARIO_CHECKS)
+    @utils.init_test_scenario(test_data.SCENARIO_CHECKS)
     @utils.create_data_root({'foo.log': '2021-04-01 00:31:00.000 an event\n',
                              'uptime': (' 16:19:19 up 17:41,  2 users, '
                                         ' load average: 3.58, 3.27, 2.58'),
@@ -638,7 +95,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
 
         self.assertEqual(IssuesManager().load_issues(), {})
 
-    @utils.init_test_scenario(SCENARIO_CHECKS)
+    @utils.init_test_scenario(test_data.SCENARIO_CHECKS)
     @utils.global_search_context
     def test_yaml_def_scenario_checks_requires(self, global_searcher):
         checker = scenarios.YScenarioChecker(global_searcher)
@@ -667,7 +124,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
 
         self.assertEqual(IssuesManager().load_issues(), {})
 
-    @utils.init_test_scenario(SCENARIO_CHECKS)
+    @utils.init_test_scenario(test_data.SCENARIO_CHECKS)
     @utils.create_data_root({'foo.log':
                              ('2021-03-29 00:31:00.000 an event\n'
                               '2021-03-30 00:32:00.000 an event\n'
@@ -757,7 +214,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
             result = ExtraSearchConstraints.filter_by_period(results, 24)
             self.assertEqual(len(result), 2)
 
-    @utils.init_test_scenario(YDEF_NESTED_LOGIC)
+    @utils.init_test_scenario(test_data.YDEF_NESTED_LOGIC)
     @utils.global_search_context
     def test_yaml_def_nested_logic(self, global_searcher):
         scenarios.YScenarioChecker(global_searcher).run()
@@ -765,7 +222,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
         self.assertEqual(sorted([issue['message'] for issue in issues]),
                          sorted(['conc1', 'conc3']))
 
-    @utils.init_test_scenario(YAML_DEF_REQUIRES_MAPPED)
+    @utils.init_test_scenario(test_data.YAML_DEF_REQUIRES_MAPPED)
     @utils.global_search_context
     def test_yaml_def_mapped_overrides(self, global_searcher):
         checker = scenarios.YScenarioChecker(global_searcher)
@@ -782,7 +239,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
     @mock.patch('hotsos.core.ycheck.engine.properties.requires.requires.log')
     @mock.patch('hotsos.core.ycheck.engine.properties.requires.common.log')
     @mock.patch('hotsos.core.ycheck.engine.properties.common.log')
-    @utils.init_test_scenario(SCENARIO_W_ERROR)
+    @utils.init_test_scenario(test_data.SCENARIO_W_ERROR)
     @utils.global_search_context
     def test_failed_scenario_caught(self, global_searcher, mock_log1,
                                     mock_log2, _mock_log3,
@@ -791,7 +248,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
 
         # Check caught exception logs
         args = ('failed to import and call property %s',
-                'tests.unit.test_ycheck_scenarios.TestProperty.i_dont_exist')
+                'tests.unit.ycheck.test_scenarios.TestProperty.i_dont_exist')
         mock_log1.exception.assert_called_with(*args)
 
         args = ('requires.%s.result raised the following',
@@ -821,7 +278,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
                        "detail")
                 self.assertEqual(issue['message'], msg)
 
-    @utils.init_test_scenario(CONFIG_SCENARIO)
+    @utils.init_test_scenario(test_data.CONFIG_SCENARIO)
     @utils.create_data_root({'test.conf': '[DEFAULT]\nkey1 = 101\n'})
     @utils.global_search_context
     def test_config_scenario_fail(self, global_searcher):
@@ -830,7 +287,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
         self.assertEqual([issue['message'] for issue in issues],
                          ['cfg is bad', 'cfg is bad2'])
 
-    @utils.init_test_scenario(CONFIG_SCENARIO)
+    @utils.init_test_scenario(test_data.CONFIG_SCENARIO)
     @utils.create_data_root({'test.conf': '[DEFAULT]\nkey1 = 102\n'})
     @utils.global_search_context
     def test_config_scenario_pass(self, global_searcher):
@@ -842,7 +299,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
     @mock.patch('hotsos.core.ycheck.scenarios.log')
     @mock.patch('hotsos.core.ycheck.engine.properties.conclusions.'
                 'ScenarioException')
-    @utils.init_test_scenario(CONCLUSION_W_INVALID_BUG_RAISES)
+    @utils.init_test_scenario(test_data.CONCLUSION_W_INVALID_BUG_RAISES)
     @utils.global_search_context
     def test_raises_w_invalid_types(self, global_searcher, mock_exc, mock_log,
                                     mock_log2):
@@ -871,7 +328,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
                        "debug mode (--debug) to get more detail")
                 self.assertEqual(issue['message'], msg)
 
-    @utils.init_test_scenario(VARS)
+    @utils.init_test_scenario(test_data.VARS)
     @utils.global_search_context
     def test_vars(self, global_searcher):
         scenarios.YScenarioChecker(global_searcher).run()
@@ -894,7 +351,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
     @mock.patch('hotsos.core.ycheck.engine.properties.requires.requires.log')
     @mock.patch('hotsos.core.ycheck.engine.properties.requires.common.log')
     @mock.patch('hotsos.core.ycheck.engine.properties.common.log')
-    @utils.init_test_scenario(LOGIC_TEST)
+    @utils.init_test_scenario(test_data.LOGIC_TEST)
     @utils.global_search_context
     def test_logical_collection_and_with_fail(self, global_searcher, mock_log1,
                                               mock_log2, _mock_log3, mock_log4,
@@ -904,7 +361,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
         expected = [
             (mock_log1,
              ('failed to import and call property %s',
-              'tests.unit.test_ycheck_scenarios.TestProperty.doesntexist'),
+              'tests.unit.ycheck.test_scenarios.TestProperty.doesntexist'),
              'exception'),
             (mock_log2,
              ('requires.%s.result raised the following', 'YPropertyVarOps'),
@@ -934,28 +391,30 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
                        "detail")
                 self.assertEqual(issue['message'], msg)
 
-    @utils.init_test_scenario(NESTED_LOGIC_TEST_NO_ISSUE)
+    @utils.init_test_scenario(test_data.NESTED_LOGIC_TEST_NO_ISSUE)
     @utils.global_search_context
     def test_logical_collection_nested_no_issue(self, global_searcher):
         scenarios.YScenarioChecker(global_searcher).run()
         issues = list(IssuesStore().load().values())
         self.assertEqual(len(issues), 0)
 
-    @utils.init_test_scenario(NESTED_LOGIC_TEST_W_ISSUE)
+    @utils.init_test_scenario(test_data.NESTED_LOGIC_TEST_W_ISSUE)
     @utils.global_search_context
     def test_logical_collection_nested_w_issue(self, global_searcher):
         scenarios.YScenarioChecker(global_searcher).run()
         issues = list(IssuesStore().load().values())
         self.assertEqual(len(issues), 1)
 
-    @utils.init_test_scenario(NESTED_LOGIC_TEST_W_ISSUE, 'myscenario')
+    @utils.init_test_scenario(test_data.NESTED_LOGIC_TEST_W_ISSUE,
+                              'myscenario')
     @utils.global_search_context
     def test_scenarios_filter_none(self, global_searcher):
         sc = scenarios.YScenarioChecker(global_searcher)
         sc.load()
         self.assertEqual([s.name for s in sc.scenarios], ['myscenario'])
 
-    @utils.init_test_scenario(NESTED_LOGIC_TEST_W_ISSUE, 'myscenario')
+    @utils.init_test_scenario(test_data.NESTED_LOGIC_TEST_W_ISSUE,
+                              'myscenario')
     @utils.global_search_context
     def test_scenarios_filter_myscenario(self, global_searcher):
         HotSOSConfig.scenario_filter = 'myplugin.scenariogroup.myscenario'
@@ -963,7 +422,8 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
         sc.load()
         self.assertEqual([s.name for s in sc.scenarios], ['myscenario'])
 
-    @utils.init_test_scenario(NESTED_LOGIC_TEST_W_ISSUE, 'myscenario')
+    @utils.init_test_scenario(test_data.NESTED_LOGIC_TEST_W_ISSUE,
+                              'myscenario')
     @utils.global_search_context
     def test_scenarios_filter_nonexistent(self, global_searcher):
         HotSOSConfig.scenario_filter = 'blahblah'
@@ -971,7 +431,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
         sc.load()
         self.assertEqual([s.name for s in sc.scenarios], [])
 
-    @utils.init_test_scenario(CONCLUSION_PRIORITY_1, 'myscenario')
+    @utils.init_test_scenario(test_data.CONCLUSION_PRIORITY_1, 'myscenario')
     @utils.global_search_context
     def test_conclusion_priority_exec_highest(self, global_searcher):
         called = []
@@ -994,7 +454,7 @@ class TestYamlScenarios(utils.BaseTestCase):  # noqa, pylint: disable=too-many-p
 
         self.assertEqual(called, ['conc3'])
 
-    @utils.init_test_scenario(CONCLUSION_PRIORITY_2, 'myscenario')
+    @utils.init_test_scenario(test_data.CONCLUSION_PRIORITY_2, 'myscenario')
     @utils.global_search_context
     def test_conclusion_priority_exec_all_same(self, global_searcher):
         called = []
